% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_InfoCrit.R
\name{AICc}
\alias{AICc}
\title{AICc}
\usage{
AICc(model)
}
\arguments{
\item{model}{an object of class DImulti}
}
\value{
A numeric value of the AICc value corresponding to the model object
}
\description{
Calculate Second-order Akaike Information Criterion for a fitted DImulti object
}
