% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_InfoCrit.R
\name{BICc}
\alias{BICc}
\title{BICc}
\usage{
BICc(model)
}
\arguments{
\item{model}{an object of class DImulti}
}
\value{
A numeric value of the BICc value corresponding to the model object
}
\description{
Calculate Second-order Bayesian Information Criterion for a fitted DImulti obect
}
