% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_fit.R
\name{DImulti}
\alias{DImulti}
\title{DImulti}
\usage{
DImulti(
  y,
  eco_func = c("NA", "NA"),
  time = c("NA", "NA"),
  unit_IDs,
  prop,
  data,
  DImodel,
  FG = NULL,
  extra_fixed = NULL,
  estimate_theta = FALSE,
  theta = 1,
  method = "REML"
)
}
\arguments{
\item{y}{If the dataset is in a wide format, this argument is a vector of k column names identifying the ecosystem function values recorded from each experimental unit in the dataset. For example, if the ecosystem function columns are labelled Y1 to Y4, then \code{y = c("Y1","Y2","Y3","Y4")}. Alternatively, the column numbers can be specified, for example, \code{y = 8:11}, where ecosystem function values are in the 8th to 11th columns. \cr If the dataset is in the long/stacked format, this argument is the column name of the response value vector, for example, \code{y = "yield"}, alternatively, the column number can be supplied, \code{y = 8}}

\item{eco_func}{A vector of size two, with the first index holding the column name containing the character or factor indicator of which response was recorded (for use when stacked/long data passed through parameter y), otherwise it is the string \code{"NA"}, case insensitive. The second index should contain a string referring to the autocorrelation structure of the responses (for use in multivariate data case), options include \code{"un"} for unstructured/general and \code{"cs"} for compound symmetry. For example, \code{ecoFunc = c("Function", "cs")}, pertaining to multivariate data in a stacked format with a compound symmetry structure, or \code{ecoFunc = c("na", "un")}, meaning either univariate or wide multivariate data with an unstructured/general variance covariance matrix}

\item{time}{A vector of size two, with the first index holding the column name containing the repeated measures identifier (i.e., indicating which time point the corresponding response was recorded at), otherwise it is the string \code{"NA"}, case insensitive. The second index should contain a string referring to the autocorrelation structure of the repeated measures, options include \code{"un"} for unstructured/general, \code{"cs"} for compound symmetry, and \code{"ar1"} for an autoregressive model of order 1 (AR(1)). For example, \code{time = c("reading", "ar1")}, pertaining to repeated measures data with multiple readings on each unit with an AR(1) structure, or \code{time = c("na", "un")}, meaning multivariate data with no repeated measures (the correlation structure will be ignored)}

\item{unit_IDs}{A vector of columns names/indices containing identifiers for the experimental units from which the observations (either multiple readings of the same response or a single reading of multiple responses) are taken. Any number of identifiers can be included but they must be in nested order, from largest to smallest, e.g., \code{unitIDs = "Plot"} or \code{unitIDs = c("Plot", "ReplicateNo.")}}

\item{prop}{A vector of S column names identifying the species proportions in each community in the dataset. For example, if the species proportions columns are labelled p1 to p4, then \code{prop = c("p1","p2","p3","p4")}. Alternatively, the column numbers can be specified, for example, \code{prop = 4:7}, where species proportions are in the 4th to 7th columns}

\item{data}{A dataframe or tibble containing all previously input columns}

\item{DImodel}{A vector of strings, referring to the interaction structure of model to be fit from the full list: \code{"STR", "ID", "FULL", "E", "AV", "ADD", "FG"}}

\item{FG}{If species are classified by g functional groups, this argument takes a string vector (of length S) of the functional group to which each species referenced in the \code{prop} argument belongs. For example, for four grassland species with two grasses and two legumes: FG could be \code{FG = c("G","G","L","L")}, where G stands for grass and L stands for legume. This argument is optional but must be supplied if the \code{"FG"} interaction structure is specified in the \code{what} argument}

\item{extra_fixed}{A formula expression for any additional fixed effect terms. For example,  \code{extraFixed = ~ Treatment} or \code{extraFixed = ~ Treatment + Density}}

\item{estimate_theta}{By default, \eqn{\theta} (the power parameter on all \code{p_i * p_j} components of each interaction variable in the model) is set equal to one. Specify \code{estimate_theta = TRUE} to include the estimation of \eqn{\theta} in the specified model.}

\item{theta}{Users may specify a value of \eqn{\theta} different than 1 to fit the DI model. Note that if \code{estimate_theta = TRUE}, then \eqn{\theta} will be estimated via maximum profile log-likelihood and the value specified for \code{theta} will be overridden. Specify a vector of positive non-zero numerical values indicating the value for the non-linear parameter of the model for each ecosystem function (in alphabetical order, use \code{sort()} to find this) present in the dataset, or a single value to be used for all. For example, \code{theta = 0.5} or \code{theta = c(1, 0.5, 1)}}

\item{method}{The string \code{"REML"} or \code{"ML"}, referring to the estimation method to be used}
}
\value{
\code{DIMulti} - a custom class object containing a list of the models fit
}
\description{
A function to fit Diversity-Interactions models to data with multiple ecosystem functions and/or multiple time points
}
\details{
<here describe DI models etc. follow DImodels example> \cr
}
\examples{
## Recreation of Dooley et al. 2015 model
data(dataBEL)
m1 <- DImulti(prop = 2:5, y = c("Y"), eco_func = c("Var", "un"),
              FG=c("G", "G", "L", "L"), extra_fixed = ~ Density,
              unit_IDs = 1, theta = 1, DImodel = c("FG"), method = "REML",
              data = dataBEL)
summary(m1)

## Repeated Measures analysis of COST data
data(dataSWE)
m2 <- DImulti(...)
}
\seealso{
\code{\link[nlme]{gls}} which this function wraps \cr \cr
\code{\link[DImodels]{DImodels}} the parent package for use with data with a single ecosystem function at a single timepoint
}
