% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_classMethods.R
\name{getVarCov.DImulti}
\alias{getVarCov.DImulti}
\title{getVarCov.DImulti}
\usage{
\method{getVarCov}{DImulti}(obj, ...)
}
\arguments{
\item{obj}{an object of class DImulti}

\item{...}{some methods for this generic function require additional arguments. None are used in
this method.}
}
\value{
A variance-covariance matrix or a named list of variance-covariance matrices.
}
\description{
Extract the variance-covariance matrix from a fitted DImulti model
}
\seealso{
\code{\link[DImodelsMulti]{DImulti}}
}
