% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_classMethods.R
\name{print.DImulti}
\alias{print.DImulti}
\title{print.DImulti}
\usage{
\method{print}{DImulti}(x, ...)
}
\arguments{
\item{x}{an object of class DImulti}

\item{...}{some methods for this generic function require additional arguments. None are used in
this method.}
}
\value{
object x
}
\description{
Print details of the fitted DI models supplied
}
\seealso{
\code{\link[base]{print}} which this function wraps.
}
