% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_classMethods.R
\name{print.summary.DImulti}
\alias{print.summary.DImulti}
\title{print.summary.DImulti}
\usage{
\method{print}{summary.DImulti}(x, verbose = FALSE, digits = .Options$digits, ...)
}
\arguments{
\item{x}{an object of class DImulti}

\item{verbose}{an optional logical value used to control the amount of output when the object is printed. Defaults to FALSE.}

\item{digits}{the number of significant digits to use when printing}

\item{...}{some methods for this generic function require additional arguments. None are used in
this method.}
}
\value{
object x
}
\description{
print the details of an object of class summary.DImulti
}
\seealso{
\code{\link[base]{print}} which this function wraps.
}
