% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_classMethods.R
\name{summary.DImulti}
\alias{summary.DImulti}
\title{summary.DImulti}
\usage{
\method{summary}{DImulti}(x, verbose = FALSE)
}
\arguments{
\item{x}{an object inheriting from class DImulti, representing a list of generalized least squared fitted linear models using the Diversity-Interactions framework}

\item{verbose}{an optional logical value used to control the amount of output when the object is printed. Defaults to FALSE.}
}
\value{
object x <plus additional components (need to list, include DI notes)>
}
\description{
Print a summary of the fitted DI models supplied
}
\seealso{
\code{\link[nlme]{summary.gls}} which this function wraps.
}
