% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImodelsMulti_HelpFile.R
\docType{package}
\name{DImodelsMulti}
\alias{DImodelsMulti}
\title{Multivariate Diversity-Interactions (DI) Models with Repeated Measures}
\description{
An add-on package to the 'DImodels' package. \cr
Intended for use with data containing multiple ecosystem function responses and/or time points from a BEF study, using the Diversity-Interactions (DI) modelling approach (Kirwan et al., 2009; Dooley et al., 2015). \cr
Data should contain a column for each species' proportion, which in total sum to one, along with multiple community level response variables (in long or wide format) and/or a community level response variable taken at multiple time points (in a long format). \cr
}
\details{
\strong{Introduction to Multivariate & Repeated Measures Diversity-Interactions Models} \cr
<insert explanation of models here>
}
\examples{
#<Insert more here>
## Load the Belgium dataset
data(dataBEL)

##Example of DImulti with unstructured error term
BELmodel <- DImulti(prop = 2:5, y = c("Y"), eco_func = c("Var", "un"),
                    FG=c("G", "G", "L", "L"), extra_fixed = ~ Density,
                    unit_IDs = 1, theta = 1, DImodel = c("FG"), method = "REML",
                    data = dataBEL)
summary(BELmodel)

}
\references{
Dooley, Á., Isbell, F., Kirwan, L., Connolly, J., Finn, J. A., & Brophy, C. (2015). Testing the effects of diversity on ecosystem multifunctionality using a multivariate model. Ecology Letters, 18(11), 1242-1251. \cr \cr
Connolly J, T Bell, T Bolger, C Brophy, T Carnus, JA Finn, L Kirwan, F Isbell, J Levine, A Lüscher, V Picasso, C Roscher, MT Sebastia, M Suter and A Weigelt (2013) An improved model to predict the effects of changing biodiversity levels on ecosystem function. Journal of Ecology, 101, 344-355.  \cr \cr
Kirwan L, J Connolly, JA Finn, C Brophy, A Lüscher, D Nyfeler and MT Sebastia (2009) Diversity-interaction modelling - estimating contributions of species identities and interactions to ecosystem function. Ecology, 90, 2032-2038.  \cr \cr
}
\seealso{
\code{\link[DImodels]{DImodels} \link[DImodels]{autoDI} \link[DImodels]{DI} \link[DImodels]{DI_data} \link{DImulti}} \cr
Example datasets: \code{\link[DImodels]{dataBEL}}, \code{\link[DImodels]{dataSWE}}, \code{\link[DImodels]{simRM}}, \code{\link[DImodels]{simMV}}, \code{\link[DImodels]{simMVRM}}
}
\author{
Laura Byrne [aut, cre], Caroline Brophy [aut], Rishabh Vishwakarma [aut], Rafael de Andrade Moral [aut] \cr \cr
Maintainer: Laura Byrne \email{byrnel54@tcd.ie}
}
