% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DImulti_classMethods.R
\name{predict.DImulti}
\alias{predict.DImulti}
\title{predict.DImulti}
\usage{
\method{predict}{DImulti}(object, newdata = NULL)
}
\arguments{
\item{object}{an object of class DImulti}

\item{newdata}{an optional dataframe containing the communities from which to predict}
}
\value{
The predictions from the supplied fitted DI models for the provided newdata, or the data used to fit the model if no newdata is supplied.
}
\description{
Predict from the DI models supplied
}
\seealso{
\code{\link[nlme]{predict.gls}} which this function wraps.
}
